pyproject.toml
[project]
name = "subman"
version = "0.2.0"
description = "A Personal Subscription Management Tool"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "openpyxl>=3.1.0",
    "pandas>=2.3.3",
    "plotly>=6.5.1",
    "pydantic>=2.12.5",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]
